refactor: improve naming for clarity, remove unused imports
