+++ src/utils/simulatePayment.ts
